# src/openapi/specification.yaml
openapi: 3.0.0
info:
  title: TypeScript Express API
  version: 1.0.0
  description: A simple API with OpenAPI specification

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

paths:
  /blackholes/{name}:
    get:
      summary: Get black hole information
      description: Retrieves information about a specific black hole by name
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          description: Name of the black hole
          example: "Sagittarius A*"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: Name of the black hole
                  mass:
                    type: string
                    description: Mass of the black hole
                  source:
                    type: string
                    description: Source URL of the information
        '404':
          description: Black hole not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  mass:
                    type: string
                  source:
                    type: string
                  error:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
  /info/datetime:
    get:
      summary: Get current date and time
      description: Returns the current server date and time along with timezone information
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  datetime:
                    type: string
                    format: date-time
                  timezone:
                    type: string

  /info/version:
    get:
      summary: Get API version information
      description: Returns the current API version and environment information
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  environment:
                    type: string
                  nodeVersion:
                    type: string